Separate Doc and DocBin tests
